mybatispluse版本调整

+++ src/main/java/com/wiesel/common/base/entity/PageResp.java
@@ -31,7 +31,7 @@ public class PageResp<T> {
 	/**
 	 * 总数
 	 */
-	private Long total;
+	private int total;
 
 	/**
 	 * 数据
